+++ ct/pairdrop.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-source <(curl -fsSL https://raw.githubusercontent.com/cjlapao/MyProxmoxVE/main/misc/build.func)
+source <(curl -fsSL https://raw.githubusercontent.com/community-scripts/ProxmoxVE/main/misc/build.func)
 # Copyright (c) 2021-2025 tteck
 # Author: tteck (tteckster)
 # License: MIT | https://github.com/community-scripts/ProxmoxVE/raw/main/LICENSE
@@ -11,7 +11,7 @@ var_cpu="${var_cpu:-1}"
 var_ram="${var_ram:-512}"
 var_disk="${var_disk:-4}"
 var_os="${var_os:-debian}"
-var_version="${var_version:-12}"
+var_version="${var_version:-13}"
 var_unprivileged="${var_unprivileged:-1}"
 
 header_info "$APP"
@@ -27,13 +27,23 @@ function update_script() {
     msg_error "No ${APP} Installation Found!"
     exit
   fi
-  msg_info "Updating $APP"
-  systemctl stop pairdrop
-  cd /opt/pairdrop || exit
-  git pull
-  npm install
-  systemctl start pairdrop
-  msg_ok "Updated $APP"
+  if check_for_gh_release "pairdrop" "schlagmichdoch/PairDrop"; then
+    msg_info "Stopping Service"
+    systemctl stop pairdrop
+    msg_ok "Stopped Service"
+
+    fetch_and_deploy_gh_release "pairdrop" "schlagmichdoch/PairDrop" "tarball"
+
+    msg_info "Configuring PairDrop"
+    cd /opt/pairdrop
+    $STD npm install
+    msg_ok "Configured PairDrop"
+
+    msg_info "Starting Service"
+    systemctl start pairdrop
+    msg_ok "Started Service"
+    msg_ok "Updated Successfully"
+  fi
   exit
 }
